Add common JSON types
Port the "standard" type hierarchy (includig Undefined) from Scala 2.

The types move inside Json object to simplify the access. The Json
object becomes a new entry point for all the functionality (types,
serialization, parsing). Specific json types could be accessed via
Json.Type notation (i.e. Json.String, Json.Null and so on). This is one
extra dot in the record but one less import compared to JsonNull,
JsonString, etc... The specific json types should rarely be used in the
code so it is an improvement.

The Json is now also type of json value (it was JsonValue in the
previous version). This also reduces amount of imports needed.

An exception was also moved inside the Json object for the same reason
of simplicity.