Provide conversion from BasicInput to ParserInput
It may be convenient to implement only a simple form of the input for
prototyping functionality. And the "extra" methods in the ParserInput
could be implemented via the primitive `statefulScan` operation.

The pre-requisite for implementation is that `M` is Monad. This is not a
huge restriction as actual JSON parser requires that `M` is Monad
anyway.