Add parser-style output bits
Add multiple bits (literals, strings, objects, arrays) that could be
used in a custom writer implementation. This aligns better with various
use-cases (for example, incremental client-driven generation).

There are no special number writers as most models and APIs store
numbers as their representation so a simple `stream.write` call is
sufficient for the purpose.

The new structure is better for testing as all the basic writers could
be covered by unit tests.