Add some readers and encoders
Doing IO and using "cool APIs" seems to be somewhat tricky. The
servlet/standard IO operates on temporary arrays. Encoders usually
operate on byte buffers. And managing "external" buffers may be cool for
some APIs but not others. Anyway, we need something and this may be a
good start for the server.