Implement new "coroutine mechanics"
Change the way closures are now passed in. The new one uses "basic"
tokens (explaining what to do) and explicit "type-correlated" callback
to indicate what should be done next. No need to have complex
client-driven encoder for capturing both operation and continuation in
the context. This makes implementor's life easier. The "suspension" call
is now trivial (this is just a call to a wrapper method on the actual
suspension token). Continuation handling is also very similar, a couple
of differences is just where things are extracted from in the pattern
matching.