Extract "RoutineExecutor" into the separate class
Prepare the module to the graceful shutdown functionality. Making it
work requires certain coordination between "request acceptor", executor
and thread manager. To properly terminate everything (and await
termination), we need to `join` all the threads. So we need to capture
those threads somewhere (in the `Module` as it is the facade). However
to start the threads we need another object. This "separate object" is
now the `RoutineExecutor`. IO operations now also interact with the
executor and not the facade.

The change also introduces flow control (control valve) that could be
used to "stop" request processing and await termination of all the
requests.