Add ability to limit number of requests in flight
Having thousands requests "in flight" (being processed) is not very
cool. Yes, we can hold these but it takes more and more resources and
just build up latency for the later requests. So it is better to quickly
abort them as soon as we know processing is not going to end in a timely
way.